Some updates (#135)
* Some Updates

- Updated RequestHeader.java
- Reworked whitelist and blacklist system (not finally in this commit)
- Updated VerificationListener
- Implemented bots.discord.pw
- Improoved SafeMessage.java

* Finished blacklist system

- Updated deprecated stugg
- Updated white and blacklist

+++ src/main/java/fun/rubicon/util/SafeMessage.java
@@ -34,8 +34,11 @@ public class SafeMessage {
     }
 
     public static void sendMessage(TextChannel textChannel, String message, int deleteTime) {
-        if (hasPermissions(textChannel))
+        if (hasPermissions(textChannel) && hasDeletePermission(textChannel))
             textChannel.sendMessage(message).queue(msg -> msg.delete().queueAfter(deleteTime, TimeUnit.SECONDS));
+        else if(hasPermissions(textChannel))
+            textChannel.sendMessage(message).queue();
+
     }
 
     public static void sendMessage(TextChannel textChannel, MessageEmbed build) {
@@ -44,8 +47,10 @@ public class SafeMessage {
     }
 
     public static void sendMessage(TextChannel textChannel, MessageEmbed build, int deleteTime) {
-        if (hasPermissions(textChannel))
+        if (hasPermissions(textChannel) && hasDeletePermission(textChannel))
             textChannel.sendMessage(build).queue(msg -> msg.delete().queueAfter(deleteTime, TimeUnit.SECONDS));
+        else if(hasPermissions(textChannel))
+            textChannel.sendMessage(build).queue();
     }
 
     public static Message sendMessageBlocking(TextChannel textChannel, String message) {
@@ -59,6 +64,12 @@ public class SafeMessage {
         if (channel.getGuild().getSelfMember().hasPermission(channel, Permission.MESSAGE_READ) && channel.getGuild().getSelfMember().hasPermission(channel, Permission.MESSAGE_WRITE))
             return true;
         channel.getGuild().getOwner().getUser().openPrivateChannel().queue(privateChannel -> privateChannel.sendMessage(EmbedUtil.error("Permission Error!", "The bot need the `MESSAGE_READ` and `MESSAGE_WRITE` permissions in the ``" + channel.getName() + "` channel to run without errors.").build()));
+        return false;
+    }
+    private static boolean hasDeletePermission(TextChannel channel){
+        if(channel.getGuild().getSelfMember().hasPermission(channel, Permission.MESSAGE_MANAGE))
+            return true;
+        channel.getGuild().getOwner().getUser().openPrivateChannel().queue(privateChannel -> privateChannel.sendMessage(EmbedUtil.error("Permission Error!", "The bot need the `MESSAGE_READ` and `MESSAGE_MANAGE` permissions in the ``" + channel.getName() + "` channel to run without errors.").build()));
         return false;
     }
